mcp-supabase-http/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sse-starlette==2.1.3
psycopg[binary]==3.2.1
python-dotenv==1.0.1
modelcontextprotocol==1.2.0
